Initial Upload and Updates
This is the first upload after the transfer from the nueue repository.
Includes Updates:
Clipboard in the admin bar under the website tab.
Deactivate button is in the stats menu.
Default tasks names.
Default 20 Ranks.
Clipboard links open in a new window.
Presets for points (set up in the options page).
Repeat shows the repeat message when clicked and then you're prompted to
click again and that repeats mastery.
Shortcode is displayed in the beginning and end of the meta boxes and
insert lightbox.
Title position is fixed in the light box.