Delete go_task_status_transition()
In order for the 'transition_post_status' hooks to work properly with
the chain order meta data, the 'cmb_validate_go_*' hooks would have
to have access to the tasks' updated post status. Unfortunately, the
CMB library we are using calls its validation hooks after the
'transition_post_status' hooks, but before 'save_post' (none of our
custom options would be saved otherwise). There's no good way of
accessing the tasks' post status pre-save, so other methods of
indicating that a task is not participating in a chain will have
to be used.