Fix #132 and add support for inverted super modifier calculations
Store items act differently than any other part of Game On, as far as
performing transactions goes. In the cost section of the Store Item
edit page, postive values indicate debits (costs) and negative values
indicate credits. If a user has a super modifier of -10%, they should
be required to pay extra and will receive less for any Store Items they
purchase. For example, prior to this change, a Store Item that costs
100 Gold and awards 100 XP would actually cost 90 Gold and 90 XP for a
user with a -10% modifier. This behavior is how Tasks are intended to
act. Store Items should instead behave like so: a 100 Gold Store Item
with a 100 XP reward would become cost 110 Gold and reward 90 XP
to a user with a super modifier -10%. The opposite would be true super
modifier values > 0%.